deployment/requirements.txt
streamlit==1.43.2
pandas==2.2.3
seaborn==0.13.2
plotly==6.0.0
matplotlib==3.10.1
pillow==11.1.0
numpy==2.2.3
scikit-learn==1.6.1